// src/types/firestore.ts
export interface City {
  id: string
  name: string
  region: string
  country: string
}

export interface Masjid {
  id: string
  name: string
  address: string
  website: string
  fajr: string
  zuhr: string
  asr: string
  maghrib: string
  isha: string
  jummah_1: string
  jummah_2: string
  timetable_pdf_url: string
}